+++ app/src/main/java/com/underthykilt/poketypes/data/pokemon/PokemonEntry.kt
@@ -1,0 +1,6 @@
+package com.underthykilt.poketypes.data.pokemon
+
+data class PokemonEntry(val id: Int, val name: String) {
+    val spriteUrl: String
+        get() = "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/$id.png"
+}